Fixes for Pending Members removing data they should not see

+++ Backend/Features/NQ/Services/PlayerService.cs
@@ -160,7 +160,7 @@ public partial class PlayerService(IServiceProvider provider) : IPlayerService
 
         var playerName = (string?)await db.ExecuteScalarAsync<string>(
             "SELECT display_name FROM public.player WHERE id = @id",
-            new { id = playerId }
+            new { id = (long)playerId }
         );
 
         if (string.IsNullOrEmpty(playerName))
